pyproject.toml
[project]
name = "realtime-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=2.8.1",
    "psutil>=7.1.3",
    "pyadl>=0.1",
    "pyinstaller>=6.17.0",
    "python-dotenv>=1.2.1",
]
